+++ BodyweightFitness/WorkoutLog/WorkoutLogGeneralViewController.swift
@@ -1,66 +1,494 @@
-import UIKit
+import SnapKit
 
-import Charts
+class TitleLabel: UILabel {
+    required init?(coder aDecoder: NSCoder) {
+        super.init(coder: aDecoder)
+        
+        self.commonInit()
+    }
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        
+        self.commonInit()
+    }
+    
+    func commonInit() {
+        self.textAlignment = .left
+        self.font = UIFont.systemFont(ofSize: 20)
+        self.textColor = UIColor.black
+    }
+}
+
+class ValueLabel: UILabel {
+    required init?(coder aDecoder: NSCoder) {
+        super.init(coder: aDecoder)
+        
+        self.commonInit()
+    }
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        
+        self.commonInit()
+    }
+    
+    func commonInit() {
+        self.textAlignment = .left
+        self.font = UIFont.systemFont(ofSize: 17)
+        self.textColor = UIColor(red: 0.47, green: 0.47, blue: 0.47, alpha: 1.00)
+    }
+}
+
+class DescriptionTextView: UITextView {
+    required init?(coder aDecoder: NSCoder) {
+        super.init(coder: aDecoder)
+        
+        self.commonInit()
+    }
+    
+    override init(frame: CGRect, textContainer: NSTextContainer?) {
+        super.init(frame: frame, textContainer: textContainer)
+        
+        self.commonInit()
+    }
+    
+    func commonInit() {
+        self.textAlignment = .left
+        self.font = UIFont.systemFont(ofSize: 17)
+        self.textColor = UIColor(red: 0.47, green: 0.47, blue: 0.47, alpha: 1.00)
+        self.isEditable = false
+        self.isSelectable = false
+        self.isScrollEnabled = false
+    }
+}
+
+class CardButton: UIButton {
+    required init?(coder aDecoder: NSCoder) {
+        super.init(coder: aDecoder)
+        
+        self.commonInit()
+    }
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        
+        self.commonInit()
+    }
+    
+    func commonInit() {
+        self.contentHorizontalAlignment = .left
+        self.titleLabel?.font = UIFont.systemFont(ofSize: 20)
+        self.setTitleColor(UIColor.primaryDark(), for: .normal)
+    }
+}
 
 class WorkoutLogGeneralViewController: UIViewController {
-    @IBOutlet weak var lineChartView: LineChartView!
+    let scrollView = UIScrollView()
+    let contentView = UIView()
     
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        let unitsSold = [20.0, 4.0, 6.0, 3.0, 12.0, 16.0]
+        self.view.backgroundColor = UIColor(red:0.96, green:0.96, blue:0.97, alpha:1.00)
         
-        setupChart()
-        setChart(unitsSold)
+        self.initializeScrollView()
+        self.initializeContent()
     }
     
-    func setupChart() {
-        lineChartView.backgroundColor = UIColor.white
-        lineChartView.chartDescription?.enabled = false
-        lineChartView.dragEnabled = false
-        lineChartView.drawGridBackgroundEnabled = false
-        lineChartView.drawMarkers = false
-        lineChartView.legend.enabled = false
-        lineChartView.animate(
-            xAxisDuration: 1.0,
-            yAxisDuration: 1.0,
-            easingOption: .easeInSine
-        )
-        
-        lineChartView.xAxis.enabled = false
-        lineChartView.xAxis.drawLabelsEnabled = false
-        lineChartView.xAxis.drawGridLinesEnabled = false
-        
-        lineChartView.leftAxis.enabled = false
-        lineChartView.leftAxis.drawLabelsEnabled = false
-        lineChartView.leftAxis.drawGridLinesEnabled = false
-        
-        lineChartView.rightAxis.enabled = false
-        lineChartView.rightAxis.drawLabelsEnabled = false
-        lineChartView.rightAxis.drawGridLinesEnabled = false
-
-    }
-    
-    func setChart(_ values: [Double]) {
-        var dataEntries: [ChartDataEntry] = []
-        
-        for i in 0..<values.count {
-            let dataEntry = ChartDataEntry(x: Double(i), y: Double(values[i]))
-            dataEntries.append(dataEntry)
+    func initializeScrollView() {
+        self.view.addSubview(self.scrollView)
+        self.scrollView.snp.makeConstraints { (make) -> Void in
+            make.edges.equalTo(view)
         }
         
-        let lineChartDataSet = LineChartDataSet(values: dataEntries, label: nil)
-        let lineChartData = LineChartData()
+        self.scrollView.addSubview(self.contentView)
+        self.contentView.snp.makeConstraints { (make) -> Void in
+            make.top.bottom.equalTo(scrollView)
+            make.left.right.equalTo(view)
+        }
+    }
+    
+    func initializeContent() {
+        let view = self.createBackgroundView()
+        let card1 = self.createStatisticsCard()
+        let card2 = self.createTodaysProgressCard()
+        let card3 = self.createWorkoutLengthHistoryCard()
+        let card4 = self.createCompletionRateHistoryCard()
+        let card5 = self.createMissedExercisesCard()
+
+        let card2Title = ValueLabel()
+        card2Title.text = "Workout Progress"
+        self.contentView.addSubview(card2Title)
         
-        lineChartDataSet.mode = .cubicBezier
-        lineChartDataSet.cubicIntensity = 0.2
-        lineChartDataSet.drawCirclesEnabled = false
-        lineChartDataSet.drawValuesEnabled = false
-        lineChartDataSet.drawHorizontalHighlightIndicatorEnabled = false
-        lineChartDataSet.colors = [UIColor.primary()]
+        let card3Title = ValueLabel()
+        card3Title.text = "Workout Length History"
+        self.contentView.addSubview(card3Title)
         
-        lineChartData.addDataSet(lineChartDataSet)
+        let card4Title = ValueLabel()
+        card4Title.text = "Completion Rate History"
+        self.contentView.addSubview(card4Title)
         
-        lineChartView.data = lineChartData
+        let card5Title = ValueLabel()
+        card5Title.text = "Missed Exercises"
+        self.contentView.addSubview(card5Title)
+        
+        view.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(contentView)
+            make.left.equalTo(contentView)
+            make.right.equalTo(contentView)
+            make.height.equalTo(100)
+        }
+        
+        card1.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(contentView).offset(8)
+            make.left.right.equalTo(contentView).inset(8)
+        }
+        
+        card2Title.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(card1.snp.bottom).offset(16)
+            make.left.right.equalTo(contentView).inset(16)
+            make.bottom.equalTo(card2.snp.top).offset(-16)
+        }
+        
+        card2.snp.makeConstraints { (make) -> Void in
+            make.left.right.equalTo(contentView).inset(8)
+        }
+        
+        card3Title.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(card2.snp.bottom).offset(16)
+            make.left.right.equalTo(contentView).inset(16)
+            make.bottom.equalTo(card3.snp.top).offset(-16)
+        }
+        
+        card3.snp.makeConstraints { (make) -> Void in
+            make.left.right.equalTo(contentView).inset(8)
+        }
+        
+        card4Title.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(card3.snp.bottom).offset(16)
+            make.left.right.equalTo(contentView).inset(16)
+            make.bottom.equalTo(card4.snp.top).offset(-16)
+        }
+        
+        card4.snp.makeConstraints { (make) -> Void in
+            make.left.right.equalTo(contentView).inset(8)
+        }
+        
+        card5Title.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(card4.snp.bottom).offset(16)
+            make.left.right.equalTo(contentView).inset(16)
+            make.bottom.equalTo(card5.snp.top).offset(-16)
+        }
+        
+        card5.snp.makeConstraints { (make) -> Void in
+            make.left.right.equalTo(contentView).inset(8)
+            make.bottom.equalTo(contentView).offset(-8)
+        }
+    }
+    
+    func createBackgroundView() -> UIView {
+        let view = UIView()
+        
+        view.backgroundColor = UIColor.primary()
+        self.contentView.addSubview(view)
+        
+        return view
+    }
+    
+    func createTitleValueView(labelText: String, valueText: String) -> UIView {
+        let view = UIView()
+        
+        let label = TitleLabel()
+        label.text = labelText
+        view.addSubview(label)
+        
+        let value = ValueLabel()
+        value.text = valueText
+        view.addSubview(value)
+        
+        label.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(view)
+            make.left.equalTo(view)
+            make.right.equalTo(view)
+        }
+        
+        value.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(label.snp.bottom).offset(8)
+            make.left.equalTo(view)
+            make.right.equalTo(view)
+            make.bottom.equalTo(view)
+        }
+        
+        return view
+    }
+    
+    func createStatisticsCard() -> CardView {
+        let card = CardView()
+        self.contentView.addSubview(card)
+        
+        let topLeftLabel = TitleLabel()
+        topLeftLabel.textAlignment = .left
+        topLeftLabel.text = "19:34"
+        card.addSubview(topLeftLabel)
+        
+        let topLeftValue = ValueLabel()
+        topLeftValue.textAlignment = .left
+        topLeftValue.text = "Start Time"
+        card.addSubview(topLeftValue)
+        
+        let topRightLabel = TitleLabel()
+        topRightLabel.textAlignment = .right
+        topRightLabel.text = "19:34"
+        
+        card.addSubview(topRightLabel)
+        
+        let topRightValue = ValueLabel()
+        topRightValue.textAlignment = .right
+        topRightValue.text = "Last Updated"
+        card.addSubview(topRightValue)
+        
+        let bottomLeftLabel = TitleLabel()
+        bottomLeftLabel.textAlignment = .left
+        bottomLeftLabel.text = "--"
+        card.addSubview(bottomLeftLabel)
+        
+        let bottomLeftValue = ValueLabel()
+        bottomLeftValue.textAlignment = .left
+        bottomLeftValue.text = "Workout Length"
+        card.addSubview(bottomLeftValue)
+        
+        let bottomRightLabel = TitleLabel()
+        bottomRightLabel.textAlignment = .right
+        bottomRightLabel.text = "1"
+        card.addSubview(bottomRightLabel)
+        
+        let bottomRightValue = ValueLabel()
+        bottomRightValue.textAlignment = .right
+        bottomRightValue.text = "1"
+        card.addSubview(bottomRightValue)
+        
+        topLeftLabel.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(card).offset(20)
+            make.leading.equalTo(card).offset(16)
+            
+            make.right.equalTo(topRightLabel.snp.left)
+            
+            make.height.equalTo(24)
+        }
+        
+        topRightLabel.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(card).offset(20)
+            make.trailing.equalTo(card).offset(-16)
+            
+            make.left.equalTo(topLeftLabel.snp.right)
+            
+            make.height.equalTo(24)
+        }
+        
+        topLeftValue.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(topLeftLabel.snp.bottom).offset(8)
+            make.leading.equalTo(card).offset(16)
+            
+            make.right.equalTo(topRightValue.snp.left)
+        }
+        
+        topRightValue.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(topRightLabel.snp.bottom).offset(8)
+            make.trailing.equalTo(card).offset(-16)
+            
+            make.left.equalTo(topLeftValue.snp.right)
+        }
+        
+        bottomLeftLabel.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(topLeftValue.snp.bottom).offset(20)
+            make.leading.equalTo(card).offset(16)
+            
+            make.right.equalTo(bottomRightLabel.snp.left)
+            
+            make.height.equalTo(24)
+        }
+        
+        bottomRightLabel.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(topRightValue.snp.bottom).offset(20)
+            make.trailing.equalTo(card).offset(-16)
+            
+            make.left.equalTo(bottomLeftLabel.snp.right)
+            
+            make.height.equalTo(24)
+        }
+        
+        bottomLeftValue.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(bottomLeftLabel.snp.bottom).offset(8)
+            make.leading.equalTo(card).offset(16)
+            
+            make.right.equalTo(bottomRightValue.snp.left)
+            
+            make.bottom.equalTo(card).offset(-20)
+        }
+        
+        bottomRightValue.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(bottomRightLabel.snp.bottom).offset(8)
+            make.trailing.equalTo(card).offset(-16)
+            
+            make.left.equalTo(bottomLeftValue.snp.right)
+            
+            make.bottom.equalTo(card).offset(-20)
+        }
+        
+        return card
+    }
+    
+    func createTodaysProgressCard() -> CardView {
+        let card = CardView()
+        self.contentView.addSubview(card)
+        
+        let topLeftLabel = TitleLabel()
+        topLeftLabel.textAlignment = .left
+        topLeftLabel.text = "1 out of 9"
+        card.addSubview(topLeftLabel)
+        
+        let topLeftValue = ValueLabel()
+        topLeftValue.textAlignment = .left
+        topLeftValue.text = "Completed Exercises"
+        card.addSubview(topLeftValue)
+        
+        let topRightLabel = TitleLabel()
+        topRightLabel.textAlignment = .right
+        topRightLabel.text = "11%"
+        
+        card.addSubview(topRightLabel)
+        
+        let topRightValue = ValueLabel()
+        topRightValue.textAlignment = .right
+        topRightValue.text = "Completion Rate"
+        card.addSubview(topRightValue)
+        
+        let stackView = UIStackView()
+        stackView.axis = .vertical
+        stackView.distribution = .fill
+        stackView.alignment = .fill
+        stackView.spacing = 0
+        stackView.translatesAutoresizingMaskIntoConstraints = false
+        card.addSubview(stackView)
+        
+        let homeBarView = HomeBarView()
+        homeBarView.categoryTitle.text = "Title"
+        homeBarView.progressView.setCompletionRate(CompletionRate(percentage: 90, label: "90%"))
+        homeBarView.progressRate.text = "90%"
+        
+        stackView.addArrangedSubview(homeBarView)
+        
+        topLeftLabel.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(card).offset(20)
+            make.leading.equalTo(card).offset(16)
+            
+            make.right.equalTo(topRightLabel.snp.left)
+            
+            make.height.equalTo(24)
+        }
+        
+        topRightLabel.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(card).offset(20)
+            make.trailing.equalTo(card).offset(-16)
+            
+            make.left.equalTo(topLeftLabel.snp.right)
+            
+            make.height.equalTo(24)
+        }
+        
+        topLeftValue.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(topLeftLabel.snp.bottom).offset(8)
+            make.leading.equalTo(card).offset(16)
+            
+            make.right.equalTo(topRightValue.snp.left)
+        }
+        
+        topRightValue.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(topRightLabel.snp.bottom).offset(8)
+            make.trailing.equalTo(card).offset(-16)
+            
+            make.left.equalTo(topLeftValue.snp.right)
+        }
+        
+        stackView.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(topLeftValue.snp.bottom).offset(16)
+            make.left.equalTo(card).offset(16)
+            make.right.equalTo(card).offset(-16)
+            make.bottom.equalTo(card).offset(-16)
+
+        }
+        
+        return card
+    }
+    
+    func createWorkoutLengthHistoryCard() -> CardView {
+        let card = CardView()
+        self.contentView.addSubview(card)
+        
+        let label = TitleLabel()
+        label.text = "12 October 2017"
+        card.addSubview(label)
+
+        label.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(card).offset(20)
+            make.left.equalTo(card).offset(16)
+            make.right.equalTo(card).offset(-16)
+            make.bottom.equalTo(card).offset(-20)
+        }
+        
+        return card
+    }
+    
+    func createCompletionRateHistoryCard() -> CardView {
+        let card = CardView()
+        self.contentView.addSubview(card)
+        
+        let label = TitleLabel()
+        label.text = "12 October 2017"
+        card.addSubview(label)
+        
+        label.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(card).offset(20)
+            make.left.equalTo(card).offset(16)
+            make.right.equalTo(card).offset(-16)
+            make.bottom.equalTo(card).offset(-20)
+        }
+        
+        return card
+    }
+    
+    func createMissedExercisesCard() -> CardView {
+        let card = CardView()
+        self.contentView.addSubview(card)
+        
+        let stackView = UIStackView()
+        stackView.axis = .vertical
+        stackView.distribution = .fill
+        stackView.alignment = .fill
+        stackView.spacing = 16
+        stackView.translatesAutoresizingMaskIntoConstraints = false
+        card.addSubview(stackView)
+        
+        stackView.addArrangedSubview(
+            createTitleValueView(labelText: "Pushup", valueText: "Test")
+        )
+        
+        stackView.addArrangedSubview(
+            createTitleValueView(labelText: "Pushup", valueText: "Test")
+        )
+        
+        stackView.addArrangedSubview(
+            createTitleValueView(labelText: "Pushup", valueText: "Test")
+        )
+        
+        stackView.snp.makeConstraints { (make) -> Void in
+            make.top.equalTo(card).offset(16)
+            make.left.equalTo(card).offset(16)
+            make.right.equalTo(card).offset(-16)
+            make.bottom.equalTo(card).offset(-16)
+        }
+        
+        return card
     }
 }
